+++ src/providers/style-service.ts
@@ -3,7 +3,8 @@ import { Injectable } from '@angular/core'
 @Injectable()
 export class StyleService {
     private readonly _colors = {
-        oldYellow: '#cfb53b'
+        oldYellow: '#cfb53b',
+        visibleLight: '#F0E0E0'
     }
 
     constructor() { }